requirements.txt
fastapi
uvicorn
python-dotenv
xai-sdk[all]==2.4.6
jinja2
google-cloud-logging